SBERDOMA-513 move DomainEntitiesAccess.utils.js to correct place - accessSchema
