+++ apps/condo/domains/billing/access/BillingAccount.js
@@ -2,8 +2,8 @@
  * Generated by `createschema billing.BillingAccount 'context:Relationship:BillingIntegrationOrganizationContext:CASCADE; importId?:Text; property:Relationship:BillingProperty:CASCADE; bindingId:Text; number:Text; unit:Text; raw:Json; meta:Json'`
  */
 
-const { canReadBillingEntity } = require('./DomainEntitiesAccess.utils')
-const { canManageBillingEntity } = require('./DomainEntitiesAccess.utils')
+const { canReadBillingEntity } = require('../utils/accessSchema')
+const { canManageBillingEntity } = require('../utils/accessSchema')
 
 async function canReadBillingAccounts ({ authentication: { item: user } }) {
     return await canReadBillingEntity(user)
