+++ apps/condo/domains/billing/access/BillingAccountMeter.js
@@ -2,8 +2,8 @@
  * Generated by `createschema billing.BillingAccountMeter 'context:Relationship:BillingIntegrationOrganizationContext:CASCADE; importId?:Text; property:Relationship:BillingProperty:CASCADE; account:Relationship:BillingAccount:CASCADE; resource:Relationship:BillingMeterResource:PROTECT; raw:Json; meta:Json'`
  */
 
-const { canReadBillingEntity } = require('./DomainEntitiesAccess.utils')
-const { canManageBillingEntity } = require('./DomainEntitiesAccess.utils')
+const { canReadBillingEntity } = require('../utils/accessSchema')
+const { canManageBillingEntity } = require('../utils/accessSchema')
 
 async function canReadBillingAccountMeters ({ authentication: { item: user } }) {
     return await canReadBillingEntity(user)
