+++ apps/condo/domains/billing/access/BillingAccountMeterReading.js
@@ -2,8 +2,8 @@
  * Generated by `createschema billing.BillingAccountMeterReading 'context:Relationship:BillingIntegrationOrganizationContext:CASCADE; importId?:Text; property:Relationship:BillingProperty:CASCADE; account:Relationship:BillingAccount:CASCADE; meter:Relationship:BillingAccountMeter:CASCADE; period:CalendarDay; value1:Integer; value2:Integer; value3:Integer; date:DateTimeUtc; raw:Json; meta:Json' --force`
  */
 
-const { canReadBillingEntity } = require('./DomainEntitiesAccess.utils')
-const { canManageBillingEntity } = require('./DomainEntitiesAccess.utils')
+const { canReadBillingEntity } = require('../utils/accessSchema')
+const { canManageBillingEntity } = require('../utils/accessSchema')
 
 async function canReadBillingAccountMeterReadings ({ authentication: { item: user } }) {
     return await canReadBillingEntity(user)
