+++ apps/condo/domains/billing/access/BillingMeterResource.js
@@ -2,8 +2,8 @@
  * Generated by `createschema billing.BillingMeterResource 'name:Text'`
  */
 
-const { canReadBillingEntity } = require('./DomainEntitiesAccess.utils')
-const { canManageBillingEntity } = require('./DomainEntitiesAccess.utils')
+const { canReadBillingEntity } = require('../utils/accessSchema')
+const { canManageBillingEntity } = require('../utils/accessSchema')
 
 async function canReadBillingMeterResources ({ authentication: { item: user } }) {
     return await canReadBillingEntity(user)
