+++ apps/condo/domains/billing/access/BillingReceipt.js
@@ -2,8 +2,8 @@
  * Generated by `createschema billing.BillingReceipt 'context:Relationship:BillingIntegrationOrganizationContext:CASCADE; importId?:Text; property:Relationship:BillingProperty:CASCADE; account:Relationship:BillingAccount:CASCADE; period:CalendarDay; raw:Json; toPay:Text; services:Json; meta:Json'`
  */
 
-const { canReadBillingEntity } = require('./DomainEntitiesAccess.utils')
-const { canManageBillingEntity } = require('./DomainEntitiesAccess.utils')
+const { canReadBillingEntity } = require('../utils/accessSchema')
+const { canManageBillingEntity } = require('../utils/accessSchema')
 
 async function canReadBillingReceipts ({ authentication: { item: user } }) {
     return await canReadBillingEntity(user)
